refactor: make getContext() async w/Promise.resolve()
